+++ packages/suite/src/storage/index.ts
@@ -11,7 +11,7 @@ import { GraphData } from '@wallet-types/graph';
 import { BuyTrade, ExchangeTrade } from 'invity-api';
 import { migrate } from './migrations';
 
-const VERSION = 19; // don't forget to add migration and CHANGELOG when changing versions!
+const VERSION = 20; // don't forget to add migration and CHANGELOG when changing versions!
 
 export interface DBWalletAccountTransaction {
     tx: WalletAccountTransaction;
@@ -169,7 +169,7 @@ const onUpgrade: OnUpgradeFunc<SuiteDBSchema> = async (db, oldVersion, newVersio
         db.createObjectStore('metadata');
     } else {
         // migrate functions
-        migrate(db, oldVersion, newVersion, transaction);
+        await migrate(db, oldVersion, newVersion, transaction);
     }
 };